+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support of multiple media types in HTTP requests
 - Add basic alternative to RI REST client on Windows
 - Add 'OnSuccess' event in RI REST client
+- Add default user-agent to RI REST client
 
 ### Changed
 - Enforce https on ABS source [#108](https://github.com/nbbrd/sdmx-dl/issues/108)